backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
beautifulsoup4==4.12.3
blis==0.7.11
bnlp-toolkit==4.0.0
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cymem==2.0.8
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
emoji==1.7.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
ftfy==6.2.0
gensim==4.3.2
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.4.0
langid==1.1.6
language_data==1.2.0
lxml==5.2.2
marisa-trie==1.1.1
MarkupSafe==2.1.5
murmurhash==1.0.10
nltk==3.8.1
numpy==1.26.4
packaging==24.0
pillow==10.3.0
preshed==3.0.9
psycopg2==2.9.9
pycountry==23.12.11
pydantic==2.7.2
pydantic_core==2.18.3
python-crfsuite==0.9.10
pytz==2024.1
regex==2024.5.15
requests==2.32.2
scipy==1.12.0
sentencepiece==0.2.0
setuptools==70.0.0
six==1.16.0
sklearn-crfsuite==0.3.6
smart-open==6.4.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.0
srsly==2.4.8
tabulate==0.9.0
thinc==8.2.3
tqdm==4.66.4
typer==0.9.4
typing_extensions==4.12.0
urllib3==2.2.1
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
